+++ packages/auto_mappr_annotation/CHANGELOG.md
@@ -1,5 +1,8 @@
 [//]: # (## Unreleased)
 
+## Unreleased
+- Add type converters, use `converters` on `AutoMappr` or `MapType`. [#119](https://github.com/netglade/auto_mappr/pull/119)
+
 ## 2.0.0-beta1
 - Adhere to netglade_analysis 4.0.0. [#111](https://github.com/netglade/auto_mappr/pull/111)
 - Add `reverse` option to `MapType`. [#115](https://github.com/netglade/auto_mappr/pull/115)